delete validations

+++ app/Http/Controllers/CarteFideliteController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Client;
 use App\Program;
+use App\Transaction;
 use App\CarteFidelite;
 use Illuminate\Http\Request;
 use App\Http\Requests\AddCardRequest;
@@ -50,25 +51,27 @@ class CarteFideliteController extends Controller
     public function create()
     {
         
-        // Générer l'identifiant commercial
-        $currentYear = date('Y');
-        $latestCard = CarteFidelite::where('commercial_ID', 'like', "CARD-$currentYear%")
-            ->orderBy('commercial_ID', 'desc')
-            ->first();
-            
-        $newCardNumber = $latestCard ? ((int)substr($latestCard->commercial_ID, -5) + 1) : 1;
-        $commercial_ID = "CARD-$currentYear-" . str_pad($newCardNumber, 5, '0', STR_PAD_LEFT);
+        // Generate the commercial ID
+    $currentYear = date('Y');
+    $latestCard = CarteFidelite::where('commercial_ID', 'like', "CARD-$currentYear%")
+        ->orderBy('commercial_ID', 'desc')
+        ->first();
+        
+    $newCardNumber = $latestCard ? ((int)substr($latestCard->commercial_ID, -5) + 1) : 1;
+    $commercial_ID = "CARD-$currentYear-" . str_pad($newCardNumber, 5, '0', STR_PAD_LEFT);
 
-        $clients = Client::all();
-        $programs = Program::where('status', 'active')->get();
+    // Fetch clients who don't have a fidelity card associated with them
+    $clientsWithoutCard = Client::whereDoesntHave('carteFidelite')->get();
 
+    // Fetch active programs
+    $programs = Program::where('status', 'active')->get();
 
-        return view('carte_fidelite.create', [
-            'title' => 'New Card',
-            'commercial_ID' => $commercial_ID, // Passage de l'identifiant commercial à la vue
-            'clients' => $clients, // Passage de la liste des clients à la vue
-            'programs' => $programs,
-        ]);
+    return view('gerantCF.create', [
+        'title' => 'New Card',
+        'commercial_ID' => $commercial_ID,
+        'clientsWithoutCard' => $clientsWithoutCard,
+        'programs' => $programs,
+    ]);
     }
 
     public function store(AddCardRequest $request)
@@ -135,8 +138,14 @@ class CarteFideliteController extends Controller
 
         public function destroy(CarteFidelite $carte)
         {
-            $carte->delete();
+            // Check if the fidelity card has associated transactions
+            $hasTransactions = Transaction::where('carte_fidelite_id', $carte->id)->exists();
+            if ($hasTransactions) {
+                return redirect()->route('carte_fidelite.index')->with('warning', 'This fidelity card has active transaction(s). Please remove the transaction(s) before deleting the fidelity card.');
+            }
 
+            // If no associated transactions, proceed with deletion
+            $carte->delete();
             return redirect()->route('carte_fidelite.index')->with('message', 'Card deleted successfully!');
         }
 }
